move KeyFilter LabelFilter into types.ts
