--- src/LabelFilter.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export enum LabelFilter {
-    Null = "\0"
-}
